Make adjustments according to senior developer recommendations ( exception handling for restTemplate, remove constructors for autowired classes, etc.)

+++ MagicWand/src/main/java/com/example/xeTraining/service/MagicWandService.java
@@ -35,18 +35,24 @@ public class MagicWandService
 
     public void updateMagicWandInfo( Long id, MagicWandCatalogue magicWand)
     {
-        MagicWandCatalogue magicWandUpdated = new MagicWandCatalogue();
-        magicWandUpdated.setId(id);
-        magicWandUpdated.setName(magicWand.getName());
-        magicWandUpdated.setDesc(magicWand.getDesc());
-        magicWandUpdated.setStock(magicWand.getStock());
-        magicWandUpdated.setAge_limit(magicWand.getAge_limit());
+        if(getMagicWandById(id) != null)
+        {
+            MagicWandCatalogue magicWandUpdated = new MagicWandCatalogue();
+            magicWandUpdated.setId(id);
+            magicWandUpdated.setName(magicWand.getName());
+            magicWandUpdated.setDesc(magicWand.getDesc());
+            magicWandUpdated.setStock(magicWand.getStock());
+            magicWandUpdated.setAge_limit(magicWand.getAge_limit());
 
-        magicWandRepo.save(magicWandUpdated);
+            magicWandRepo.save(magicWandUpdated);
+        }
     }
     
     public void deleteMagicWandInfo(Long id)
     {
-        magicWandRepo.deleteById(id);
+        if(getMagicWandById(id) != null)
+        {
+            magicWandRepo.deleteById(id);
+        }
     }
 }
